% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaney.neighbours.r
\name{delaney.neighbours}
\alias{delaney.neighbours}
\title{delaney.neighbours}
\usage{
delaney.neighbours(points = NULL, x = NULL, y = NULL)
}
\arguments{
\item{points}{a data structure that has fields \code{points$x} and \code{points$y} with pairs 
of coordinates of the points; if one or both are absent, or the parameter in \code{NULL} (default),
the coordinates as to be passed as \code{x} and \code{y} coordinate vectors}

\item{x, y}{coordinate vectors}
}
\value{
\code{data.frame}, each row is a pair of indices of neighbouring points
}
\description{
Delaney triangulation-based pairs of neighbours
}
