% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.neighbours.r
\name{delaunay.neighbours}
\alias{delaunay.neighbours}
\alias{delaunay.neighbours.numeric}
\alias{delaunay.neighbours.SpatialExperiment}
\alias{delaunay.neighbours.default}
\title{delaunay.neighbours: create a list of neighbouring point pairs form the point coordinates}
\usage{
delaunay.neighbours(x, ...)

\method{delaunay.neighbours}{numeric}(x, y, ...)

\method{delaunay.neighbours}{SpatialExperiment}(x, ...)

\method{delaunay.neighbours}{default}(x, x.name = "x", y.name = "y", ...)
}
\arguments{
\item{x}{could be a a numeric coordinate vector. 
In this case, \code{y} parameter is required. 

If \code{x} is a \code{SpatialExperiment}, the point (cells) coordinates are read from 
\code{Xcoord} and \code{Ycood} slots.

Alternatively, \code{x} is a 
data structure that can be subsetted by \code{$x} and \code{$y} 
or by the names provided by \code{x.name} and \code{y.name}) parameters to get the point coordinates.}

\item{y}{second coordinate vector}

\item{x.name}{the field of column name to subset \code{x$x.name}, the default is "x"}

\item{y.name}{the field of column name to subset \code{x$y.name}, the default is "y"}
}
\value{
\code{data.frame}, each row is a pair of indices of neighbouring points
}
\description{
delaunay.neighbours
}
\details{
Delaunay triangulation-based pairs of neighbours
}
\examples{
x <- runif(50)
y <- runif(50)
df<-data.frame(x=x,y=y)
neighb.xy<-delaunay.neighbours(x,y)
neighb.df<-delaunay.neighbours(df)
dfc<-data.frame(x.coord=x,y.coord=y)
neighb.dfc<-delaunay.neighbours(dfc,x.name="x.coord",y.name="y.coord")
}
