% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.neighbours.r
\name{delaunay.neighbours}
\alias{delaunay.neighbours}
\alias{delaunay.neighbours.numeric}
\alias{delaunay.neighbours.default}
\title{delaunay.neighbours: create a list of neighbouring point pairs form the point coordinates}
\usage{
delaunay.neighbours(x, ...)

\method{delaunay.neighbours}{numeric}(x, y, ...)

\method{delaunay.neighbours}{default}(x, x.name = "x", y.name = "y", ...)
}
\arguments{
\item{x}{could be a data structure that can be subsetted by \code{$x} and \code{$y} 
(the names can be altered by parameters #' \code{x.name} and \code{y.name}) or a numeric coordinate vector. In this case, y parameter is required.}

\item{y}{second coordinate vector}

\item{x.name}{the field of column name to subset \code{x$x.name}, the default is "x"}

\item{y.name}{the field of column name to subset \code{x$y.name}, the default is "y"}
}
\value{
\code{data.frame}, each row is a pair of indices of neighbouring points
}
\description{
delaunay.neighbours
}
\details{
Delaunay triangulation-based pairs of neighbours
}
